move internal/codeintel/lsif/existence => lib/codeintel/pathexistence
