+++ enterprise/lib/codeintel/pathexistence/directory_contents_test.go
@@ -1,4 +1,4 @@
-package existence
+package pathexistence
 
 import (
 	"context"
