+++ enterprise/lib/codeintel/pathexistence/existence_checker.go
@@ -1,4 +1,4 @@
-package existence
+package pathexistence
 
 import (
 	"context"
