+++ enterprise/lib/codeintel/pathexistence/paths_test.go
@@ -1,4 +1,4 @@
-package existence
+package pathexistence
 
 import (
 	"testing"
